+++ server/api/v1/apikey/[id].put.mjs
@@ -7,7 +7,7 @@ export default defineEventHandler(async (event) => {
     const { id } = getRouterParams(event);
     const apikey = await APIKeys.findOrDie(id);
     const session = obtainSession(event);
-    apikey.canEdit(session);
+    await apikey.canEdit(session);
     await apikey.updateAndVerify(await getBody(event), session);
     return apikey.describe(describeParams(event));
 });