+++ server/api/v1/s3file/[id].delete.mjs
@@ -5,7 +5,7 @@ export default defineEventHandler(async (event) => {
     const S3Files = await useKind('S3File');
     const { id } = getRouterParams(event);
     const s3file = await S3Files.findOrDie(id);
-    s3file.canEdit(obtainSession(event));
+    await s3file.canEdit(obtainSession(event));
     await s3file.delete();
     return s3file.describe(describeParams(event));
 });