+++ server/api/v1/s3file/[id].put.mjs
@@ -6,7 +6,7 @@ export default defineEventHandler(async (event) => {
     const { id } = getRouterParams(event);
     const s3file = await S3Files.findOrDie(id);
     const session = obtainSession(event);
-    s3file.canEdit(session);
+    await s3file.canEdit(session);
     await s3file.updateAndVerify(await getBody(event), session);
     return s3file.describe(describeParams(event));
 });