+++ docs/change-log.md
@@ -3,6 +3,9 @@ outline: deep
 ---
 # Change Log
 
+## 2024-04-08
+* Breaking change: Refactored VingRecord isOwner(), canEdit(), and propsOptions() to be async.
+
 ## 2024-04-07
 * Made the MenuBar in the default layout client only until PrimeVue fixes #5541.
 * useVingKind().mint() now carries forward the query params into the new record.